// react-module/src/openapi/models/contractor-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ContractorDto
 */
export interface ContractorDto {
    /**
     * 
     * @type {number}
     * @memberof ContractorDto
     */
    id?: any;
    /**
     * 
     * @type {Gid}
     * @memberof ContractorDto
     */
    gid?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorDto
     */
    acronym?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorDto
     */
    name1?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorDto
     */
    postCode?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorDto
     */
    city?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorDto
     */
    street?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorDto
     */
    nip?: any;
    /**
     * 
     * @type {string}
     * @memberof ContractorDto
     */
    personalIdNumber?: any;
}